Landing Target Update
Expanded the environment to support a configurable landing target for the agent, and expanded sensors to account for this. An initial network was trained, but further work required.
Also recalibrated / reset component positions, as were previously offset from world origin, and updated position requests to use Local Positions (relative to parent)