Extended run with curriculum learning
Ran for 1e8 steps, with curriculum learning (and finite lifetime = 500 steps). Capable of getting to the target sometimes, however with weird flip manoeuvre. Consider revising the reward function, modifying the lifetime (so can explore hovering as before), and / or making the environment easier by removing initial orientation randomisation and / or reducing the max 2D distance between the rocket and target.