New Agent Trained, Epsilon=0.4
Trained a new agent using epsilon=0.4, which appeared to perform much better with an identical environment to before. Rationale for the choice of epsilon was to increase the amount of divergence between policies to try and cover more of the policy space and avoid settling into as much of a local optimum like the last time, where it adopted a flipping behaviour which was undesirable.