Environment simplification
Default / base target size increased to 5.0f, and max distance from target to rocket decreased (enforced to be max of 30f). This reduces the likeliness of very challenging scenarios that the agent is likely to fail and struggle to learn from, as evidenced in the last training cycle. Corresponding change made in Model_Configurator.py curriculum.