Matplotlib Debugging
Added Matplotlib plot to debug the data. Concluded that there is an error in the data formatting, likely as a result of ignoring inconsistent data output frequency for the different scalars. Next steps could include accounting for all the unique steps in the initial array creation (and sizing accordingly), creating a set of lists, which could be accessed per loop depending on current step of interest, or other - to be investigated.