Rocket Training Update
Agent trained using simplified environment. Larger target appears to have decreased the accuracy, and the rocket has a greater tendency to enter a flip. Perhaps look at reducing the target size again, and / or changing the way the curriculum progresses, and consider adding a reward for stability (e.g. a negative penalty when inverted, or something relating to the angular velocity)