16 Agent Test
Completed a training run with 16 agents, which took 10 hours. The agent appeared to learn how to land the rocket within the target (i.e. did better than the 64 agent run, however appears slightly less precise and accurate than the 4 agent case). The relationship between number of agents and time to train can be modelled by a parametric equation, with N being an increasing integer, the number of agents being 4^N and the time to train being 40/(8^N). However, while increasing the number of agents appears to speed up training time, this does not mean the agent is trained to the same standard. This is because while the same number of steps are assessed over the a shorter time period (across all the environments), these environments show less variation, and as such, the neural networks are updated over a less distributed spread of actions and observations, limiting exploration of the policy space and hence producing a worse outcome if increased indefinitely.